Promote tournaments.py to a package split by sub-topic
Converts app/routes/tournaments.py into a package with three topical
submodules: management (TO-side CRUD, fields, tags), scheduling
(schedule editing, recompute, import/export, autocomplete, DSL
validation), and recordings (cameras, recording/preview endpoints,
ffmpeg finalisation, user-upload pipeline). All 39 routes share the
same Blueprint object defined in __init__.py, so URL endpoint names
are unchanged. _api.py is not touched.

Adds a 'When to promote a blueprint to a package' section to
app/routes/README.md documenting the trigger conditions (~1,000 lines,
3+ distinct sub-topics, or recurring merge conflicts) and the
mechanical conversion steps.
