+++ app/routes/tournaments/management.py
@@ -1,0 +1,739 @@
+"""TO-side tournament and league management routes.
+
+Create / edit / delete tournaments and leagues, field CRUD, tag CRUD,
+TO membership management. Part of the ``tournaments`` blueprint.
+"""
+
+from flask import (
+    Blueprint,
+    request,
+    jsonify,
+    current_app,
+)
+from flask_login import login_required, current_user
+from flask_executor import Executor
+
+from datetime import datetime, timedelta, timezone
+import json
+import time
+import uuid
+
+from models import (
+    Tournament,
+    Match,
+    Field,
+    Tag,
+    Camera,
+    Point,
+    TeamRegistration,
+    PlayerRegistration,
+    Team,
+    TO,
+    League,
+    db,
+)
+from app.services._common import current_user_type
+from app.utils.helpers import (
+    resolve_team_name_to_id,
+    resolve_tag_to_team,
+)
+from app.utils.scheduling import (
+    compute_dynamic_match_nominal_start_time,
+    validate_match_input,
+    recompute_all_match_times,
+)
+from app.utils.name_validation import match_name_char_error
+from app.models.validators import URL_SLUG_ALLOWED_HINT, is_valid_url_slug
+from app.utils.decorators import require_tournament_organizer
+from app.utils.datetime_helpers import now_utc_naive
+
+from os import path, listdir
+
+from app.utils.footage import finalize_recording_worker
+from app.utils.user_uploads import (
+    create_direct_user_upload_camera,
+    list_batch_manifest_rows,
+    register_batch_upload_completion,
+)
+from app.utils.camera_helpers import (
+    generate_camera_key,
+    require_camera_key,
+)
+from app.utils.recording_retry import (
+    RETRY_FINALIZATION_USER_IDS_ENV,
+    current_user_can_retry_finalization,
+)
+from app.utils import preview_store
+
+from app.domain.enums import (
+    MatchStatus,
+    ScheduleType,
+    SetType,
+)
+from app.services.registration_resolver import is_player_registered
+from app.services.permission_service import PermissionService
+
+from . import bp
+
+
+@bp.route("/create-tournament", methods=["POST"])
+@login_required
+def create_tournament():
+    """Create a new tournament and assign the creator as TO.
+
+    ``POST /_api/create-tournament``
+
+    Creates the tournament record and, for standalone tournaments, a
+    :class:`~app.models.registrable_config.RegistrableConfig`.  When
+    *league_id* is provided the tournament inherits the league's config and
+    the caller must be a league TO.
+
+    Form Data:
+        name (str): Display name for the tournament.
+        url (str): URL slug (must be unique).
+        league_id (str | None): Optional league to attach to.
+
+    Returns:
+        JSON ``{"success": true, "url": "<slug>"}`` on success, or error
+        with HTTP 400/403.
+    """
+    name = request.form["name"].strip()
+    url = request.form["url"].strip()
+
+    if not name or not url:
+        return (
+            jsonify(
+                {
+                    "success": False,
+                    "error": f"Name and URL slug are required. {URL_SLUG_ALLOWED_HINT}",
+                }
+            ),
+            400,
+        )
+    if not is_valid_url_slug(url):
+        return jsonify({"success": False, "error": URL_SLUG_ALLOWED_HINT}), 400
+
+    if Tournament.query.filter_by(url=url).first():
+        return (
+            jsonify({"success": False, "error": "Tournament URL already exists"}),
+            400,
+        )
+
+    league_id = None
+    raw_league_id = request.form.get("league_id", "").strip()
+    if raw_league_id:
+        league = League.query.get(raw_league_id)
+        if not league:
+            return jsonify({"success": False, "error": "League not found"}), 400
+        if not PermissionService.is_league_organizer(raw_league_id, current_user):
+            return (
+                jsonify(
+                    {
+                        "success": False,
+                        "error": "You must be an organizer of that league to attach a tournament to it.",
+                    }
+                ),
+                403,
+            )
+        league_id = raw_league_id
+
+    from models import RegistrableConfig
+
+    start_date = now_utc_naive()
+    tournament = Tournament(
+        url=url,
+        name=name,
+        start_date=start_date,
+        end_date=start_date,
+        league_id=league_id,
+    )
+    if not league_id:
+        rc = RegistrableConfig(
+            team_reg_fee=0.0,
+            player_reg_fee=0.0,
+        )
+        db.session.add(rc)
+        db.session.flush()
+        tournament.registrable_config_id = rc.id
+
+    db.session.add(tournament)
+    db.session.flush()
+
+    to_entry = TO(
+        user_id=current_user.id,
+        user_type=current_user_type(),
+        event=url,
+    )
+    db.session.add(to_entry)
+    db.session.commit()
+
+    return (
+        jsonify(
+            {
+                "success": True,
+                "message": f'Tournament "{name}" created successfully!',
+                "url": url,
+            }
+        ),
+        200,
+    )
+
+
+@bp.route("/create-league", methods=["POST"])
+@login_required
+def create_league():
+    """Create a new league. TOs create a new league for each season."""
+    from models import League, TO, db
+
+    league_name = request.form.get("league_name", "").strip()
+    league_url = request.form.get("league_url", "").strip()
+
+    if not league_name or not league_url:
+        return (
+            jsonify(
+                {
+                    "success": False,
+                    "error": f"League name and URL slug are required. {URL_SLUG_ALLOWED_HINT}",
+                }
+            ),
+            400,
+        )
+    if not is_valid_url_slug(league_url):
+        return jsonify({"success": False, "error": URL_SLUG_ALLOWED_HINT}), 400
+
+    if League.query.filter_by(url=league_url).first():
+        return jsonify({"success": False, "error": "League URL already exists"}), 400
+
+    from models import RegistrableConfig
+
+    rc = RegistrableConfig(
+        team_reg_fee=0.0,
+        player_reg_fee=0.0,
+    )
+    db.session.add(rc)
+    db.session.flush()
+    league = League(
+        url=league_url,
+        name=league_name,
+        registrable_config_id=rc.id,
+    )
+    db.session.add(league)
+    db.session.flush()
+
+    to_entry = TO(
+        user_id=current_user.id,
+        user_type=current_user_type(),
+        event=None,
+        league_id=league_url,
+    )
+    db.session.add(to_entry)
+    db.session.commit()
+
+    return (
+        jsonify(
+            {
+                "success": True,
+                "message": f'League "{league_name}" created successfully!',
+                "league_url": league_url,
+            }
+        ),
+        200,
+    )
+
+
+@bp.route("/<tournament_url>/update-settings", methods=["POST"])
+@require_tournament_organizer("Only tournament organizers can access this page")
+def update_tournament_settings(tournament_url):
+    """Update tournament settings."""
+    tournament = Tournament.query.filter_by(url=tournament_url).first_or_404()
+
+    tournament.name = request.form["name"]
+    tournament.location = request.form.get("location", "")
+    tournament.about = request.form.get("about", "")
+    from app.services.dual_write import set_head_ref_allowlist_from_csv
+
+    set_head_ref_allowlist_from_csv(tournament, request.form.get("head_refs_allowed_list", ""))
+    tournament.head_refs_allow_reffing_teams = "head_refs_allow_reffing_teams" in request.form
+    tournament.head_refs_allow_anyone = "head_refs_allow_anyone" in request.form
+    tournament.published = "published" in request.form
+    tournament.schedule_published = "schedule_published" in request.form
+    if not tournament.league_id and tournament.registrable_config:
+        rc = tournament.registrable_config
+        rc.team_reg_fee = float(request.form.get("team_reg_fee", 0))
+        rc.player_reg_fee = float(request.form.get("player_reg_fee", 0))
+        rc.terms_link = request.form.get("terms_link", "") or None
+        rc.team_registration_open = "team_registration_open" in request.form
+        rc.player_registration_open = "player_registration_open" in request.form
+        n_max = request.form.get("n_max_teams", "").strip()
+        rc.n_max_teams = int(n_max) if n_max else None
+        roster = request.form.get("max_team_size_roster", "").strip()
+        rc.max_team_size_roster = int(roster) if roster else None
+        field = request.form.get("max_team_size_field", "").strip()
+        rc.max_team_size_field = int(field) if field else None
+        if "require_waiver_signature" not in request.form:
+            rc.waiver_filepath = None
+            rc.waiver_sha256 = None
+
+    if request.form.get("start_date"):
+        tournament.start_date = datetime.strptime(request.form["start_date"], "%Y-%m-%d")
+
+    end_date_val = request.form.get("end_date", "").strip()
+    if not end_date_val:
+        return (
+            jsonify(
+                {
+                    "success": False,
+                    "error": "End date is required.",
+                }
+            ),
+            400,
+        )
+    tournament.end_date = datetime.strptime(end_date_val, "%Y-%m-%d")
+
+    db.session.commit()
+    return (
+        jsonify({"success": True, "message": "Tournament settings updated successfully!"}),
+        200,
+    )
+
+
+@bp.route("/<tournament_url>/delete", methods=["POST"])
+@require_tournament_organizer("Only tournament organizers can access this page")
+def delete_tournament(tournament_url):
+    """Delete a tournament and all related data."""
+    tournament = Tournament.query.filter_by(url=tournament_url).first_or_404()
+
+    # Verify confirmation URL slug
+    confirm_url = request.form.get("confirm_url", "").strip()
+    if confirm_url != tournament_url:
+        return (
+            jsonify(
+                {
+                    "success": False,
+                    "error": "Confirmation URL does not match. Tournament not deleted.",
+                }
+            ),
+            400,
+        )
+
+    # Import all necessary models
+    from models import (
+        Point,
+        MatchNote,
+        Match,
+        HeadRef,
+        Field,
+        Tag,
+        SideComp,
+        SideCompRegistration,
+        SideCompResult,
+        PenaltyType,
+    )
+
+    # Delete in order to respect foreign key constraints.
+    # Order: side comp results & registrations -> side comps; points & match notes -> matches;
+    # then penalty types, head refs, registrations, TOs, fields, tags; finally tournament.
+
+    side_comps = SideComp.query.filter_by(event=tournament_url).all()
+    side_comp_ids = [sc.id for sc in side_comps]
+    if side_comp_ids:
+        SideCompResult.query.filter(SideCompResult.comp.in_(side_comp_ids)).delete(synchronize_session=False)
+        SideCompRegistration.query.filter(SideCompRegistration.comp.in_(side_comp_ids)).delete(
+            synchronize_session=False
+        )
+
+    SideComp.query.filter_by(event=tournament_url).delete(synchronize_session=False)
+
+    matches = Match.query.filter_by(event=tournament_url).all()
+    match_uuids = [m.uuid for m in matches]
+    if match_uuids:
+        Point.query.filter(Point.match.in_(match_uuids)).delete(synchronize_session=False)
+        MatchNote.query.filter(MatchNote.match.in_(match_uuids)).delete(synchronize_session=False)
+    Match.query.filter_by(event=tournament_url).delete(synchronize_session=False)
+    # PenaltyType after MatchNote (notes reference penalty_type_id)
+    # Only delete event-level penalty types; league events use league's penalty types
+    if not tournament.league_id:
+        PenaltyType.query.filter_by(event=tournament_url).delete(synchronize_session=False)
+    HeadRef.query.filter_by(event=tournament_url).delete(synchronize_session=False)
+    PlayerRegistration.query.filter_by(event=tournament_url).delete(synchronize_session=False)
+    TeamRegistration.query.filter_by(event=tournament_url).delete(synchronize_session=False)
+    Field.query.filter_by(event=tournament_url).delete(synchronize_session=False)
+    Tag.query.filter_by(event=tournament_url).delete(synchronize_session=False)
+    TO.query.filter_by(event=tournament_url).delete(synchronize_session=False)
+    rc_id = tournament.registrable_config_id if not tournament.league_id else None
+    db.session.delete(tournament)
+    if rc_id:
+        from models import RegistrableConfig
+
+        rc = RegistrableConfig.query.get(rc_id)
+        if rc:
+            db.session.delete(rc)
+    db.session.commit()
+
+    return (
+        jsonify(
+            {
+                "success": True,
+                "message": f'Tournament "{tournament.name}" has been permanently deleted.',
+            }
+        ),
+        200,
+    )
+
+
+@bp.route("/<tournament_url>/add-to", methods=["POST"])
+@login_required
+def add_to(tournament_url):
+    """Add a TO to the tournament."""
+
+    tournament = Tournament.query.filter_by(url=tournament_url).first_or_404()
+    if tournament.league_id:
+        return (
+            jsonify(
+                {
+                    "success": False,
+                    "error": "TOs for league events are managed from the league page.",
+                }
+            ),
+            403,
+        )
+
+    if not PermissionService.is_tournament_organizer(tournament_url, current_user):
+        return (
+            jsonify(
+                {
+                    "success": False,
+                    "error": "Only tournament organizers can access this page",
+                }
+            ),
+            403,
+        )
+
+    user_id = request.form.get("user_id", "").strip()
+    user_type = request.form.get("user_type", "").strip().lower()
+
+    if not user_id or user_type not in ["player", "team"]:
+        return jsonify({"success": False, "error": "Invalid user ID or type"}), 400
+
+    # Verify the user exists
+    from models import Player
+
+    if user_type == "player":
+        user = Player.query.get(user_id)
+        if not user:
+            return (
+                jsonify({"success": False, "error": f'Player with ID "{user_id}" not found'}),
+                404,
+            )
+    else:  # team
+        user = Team.query.get(user_id)
+        if not user:
+            return (
+                jsonify({"success": False, "error": f'Team with ID "{user_id}" not found'}),
+                404,
+            )
+
+    # Check if TO already exists
+    existing_to = TO.query.filter_by(user_id=user_id, user_type=user_type, event=tournament_url).first()
+
+    if existing_to:
+        return (
+            jsonify(
+                {
+                    "success": False,
+                    "error": "This user is already a TO for this tournament",
+                }
+            ),
+            400,
+        )
+
+    # Create new TO entry
+    new_to = TO(user_id=user_id, user_type=user_type, event=tournament_url)
+    db.session.add(new_to)
+    db.session.commit()
+
+    user_name = user.name if user else user_id
+    return (
+        jsonify({"success": True, "message": f"Successfully added {user_name} as a TO"}),
+        200,
+    )
+
+
+@bp.route("/<tournament_url>/remove-to", methods=["POST"])
+@login_required
+def remove_to(tournament_url):
+    """Remove a TO from the tournament."""
+
+    tournament = Tournament.query.filter_by(url=tournament_url).first_or_404()
+    if tournament.league_id:
+        return (
+            jsonify(
+                {
+                    "success": False,
+                    "error": "TOs for league events are managed from the league page.",
+                }
+            ),
+            403,
+        )
+
+    if not PermissionService.is_tournament_organizer(tournament_url, current_user):
+        return (
+            jsonify(
+                {
+                    "success": False,
+                    "error": "Only tournament organizers can access this page",
+                }
+            ),
+            403,
+        )
+
+    to_id = request.form.get("to_id")
+    if not to_id:
+        return jsonify({"success": False, "error": "TO ID is required"}), 400
+
+    # Get the TO entry to remove
+    to_to_remove = TO.query.get_or_404(to_id)
+
+    # Verify it's for this tournament
+    if to_to_remove.event != tournament_url:
+        return jsonify({"success": False, "error": "Invalid TO entry"}), 400
+
+    # Prevent removing yourself (optional - you might want to allow this)
+    if to_to_remove.user_id == current_user.id and to_to_remove.user_type == current_user_type():
+        return (
+            jsonify({"success": False, "error": "You cannot remove yourself as a TO"}),
+            400,
+        )
+
+    # Get user info for flash message
+    from models import Player
+
+    if to_to_remove.user_type == "player":
+        user = Player.query.get(to_to_remove.user_id)
+    else:
+        user = Team.query.get(to_to_remove.user_id)
+    user_name = user.name if user else to_to_remove.user_id
+
+    # Delete the TO entry
+    db.session.delete(to_to_remove)
+    db.session.commit()
+
+    return (
+        jsonify({"success": True, "message": f"Successfully removed {user_name} as a TO"}),
+        200,
+    )
+@bp.route("/<tournament_url>/add-field", methods=["POST"])
+@require_tournament_organizer("Only tournament organizers can access this page")
+def add_field(tournament_url):
+    """Add a field to tournament."""
+    tournament = Tournament.query.filter_by(url=tournament_url).first_or_404()
+
+    # Get camera URLs from form (camera[] array)
+    camera_urls = request.form.getlist("camera[]")
+    # Filter out empty values
+    camera_urls = [url.strip() for url in camera_urls if url.strip()]
+
+    # Store as JSON array
+    camera_value = json.dumps(camera_urls) if camera_urls else ""
+
+    field = Field(event=tournament_url, name=request.form["field_name"], camera=camera_value)
+
+    db.session.add(field)
+    db.session.commit()
+
+    return jsonify({"success": True, "message": "Field added successfully!"}), 200
+
+
+@bp.route("/<tournament_url>/update-field", methods=["POST"])
+@require_tournament_organizer("Only tournament organizers can access this page")
+def update_field(tournament_url):
+    """Update field."""
+    field_id = request.form.get("field_id")
+    if not field_id:
+        return jsonify({"success": False, "error": "Field ID is required"}), 400
+
+    field = Field.query.get_or_404(field_id)
+    old_field_name = field.name
+    new_field_name = request.form["field_name"]
+
+    # Update field name
+    field.name = new_field_name
+
+    # Get camera URLs from form (camera[] array)
+    camera_urls = request.form.getlist("camera[]")
+    # Filter out empty values
+    camera_urls = [url.strip() for url in camera_urls if url.strip()]
+
+    # Get old camera URLs for comparison
+    old_camera_urls = []
+    if field.camera:
+        from app.utils.camera_helpers import parse_camera_urls
+
+        old_camera_urls = parse_camera_urls(field.camera)
+
+    # Store as JSON array
+    field.camera = json.dumps(camera_urls) if camera_urls else ""
+
+    # Get all matches that reference this field (for both name and camera updates)
+    # Use old field name if name changed, otherwise use current name
+    field_name_for_query = old_field_name if old_field_name != new_field_name else new_field_name
+    matches_to_update = Match.query.filter_by(event=tournament_url, field=field_name_for_query).all()
+
+    # If camera URLs changed, update matches and points that reference this field
+    camera_urls_changed = old_camera_urls != camera_urls
+    camera_update_count = 0
+    if camera_urls_changed:
+        # Build mapping from old index to new index based on URL matching
+        # This handles reordering, additions, and removals
+        old_to_new_index_map = {}
+        for new_idx, new_url in enumerate(camera_urls):
+            # Find if this URL existed in old list
+            try:
+                old_idx = old_camera_urls.index(new_url)
+                old_to_new_index_map[str(old_idx)] = str(new_idx)
+            except ValueError:
+                # New URL, no mapping needed
+                pass
+
+        # Update matches that reference this field
+        for match in matches_to_update:
+            if match.camera_stream_starts:
+                try:
+                    stream_starts = json.loads(match.camera_stream_starts)
+                    # Remap camera indices
+                    new_stream_starts = {}
+                    for old_idx_str, start_time in stream_starts.items():
+                        if old_idx_str in old_to_new_index_map:
+                            new_idx_str = old_to_new_index_map[old_idx_str]
+                            new_stream_starts[new_idx_str] = start_time
+                        # If old index not in map, camera was removed - don't include it
+                    match.camera_stream_starts = json.dumps(new_stream_starts) if new_stream_starts else None
+                    camera_update_count += 1
+                except (json.JSONDecodeError, TypeError) as e:
+                    print(f"Error updating camera_stream_starts for match {match.uuid}: {e}")
+                    # If parsing fails, clear it
+                    match.camera_stream_starts = None
+
+        # Update points that reference this field (via the match)
+        # Get all points for matches on this field
+        from models import Point
+        from app.utils.camera_helpers import calculate_stream_timestamp
+
+        point_update_count = 0
+        for match in matches_to_update:
+            points = Point.query.filter_by(match=match.uuid).all()
+
+            # Get stream start times for this match
+            stream_starts = {}
+            if match.camera_stream_starts:
+                try:
+                    stream_starts = json.loads(match.camera_stream_starts)
+                except (json.JSONDecodeError, TypeError):
+                    pass
+
+            for point in points:
+                # First, handle camera_index remapping if needed
+                if point.camera_index is not None:
+                    old_idx_str = str(point.camera_index)
+                    if old_idx_str in old_to_new_index_map:
+                        # Remap to new index
+                        new_idx = int(old_to_new_index_map[old_idx_str])
+                        point.camera_index = new_idx
+                        point_update_count += 1
+                    else:
+                        # Camera at this index was removed - try to find matching URL
+                        # If we can't find it, set to None
+                        if point.camera_index < len(old_camera_urls):
+                            old_url = old_camera_urls[point.camera_index]
+                            try:
+                                new_idx = camera_urls.index(old_url)
+                                point.camera_index = new_idx
+                                point_update_count += 1
+                            except ValueError:
+                                # URL not found in new list, set to None
+                                point.camera_index = None
+                                point.stream_timestamp = None
+                                point_update_count += 1
+                        else:
+                            # Index was out of bounds, set to None
+                            point.camera_index = None
+                            point.stream_timestamp = None
+                            point_update_count += 1
+
+                # Recompute stream_timestamp for all points that have a camera_index and stamp
+                # This ensures timestamps are recalculated based on current stream start times
+                if point.camera_index is not None and point.stamp:
+                    camera_idx_str = str(point.camera_index)
+                    if camera_idx_str in stream_starts:
+                        stream_start_time = stream_starts[camera_idx_str]
+                        new_timestamp = calculate_stream_timestamp(point.stamp, stream_start_time)
+                        if new_timestamp is not None:
+                            point.stream_timestamp = new_timestamp
+                            point_update_count += 1
+
+    # Propagate field name change to all matches that reference this field
+    name_update_count = 0
+    if old_field_name != new_field_name:
+        for match in matches_to_update:
+            match.field = new_field_name
+            name_update_count += 1
+
+    # Generate success message
+    update_messages = []
+    if name_update_count > 0:
+        update_messages.append(f"Updated {name_update_count} match(es) to use the new field name")
+    if camera_urls_changed:
+        if camera_update_count > 0:
+            update_messages.append(f"Updated camera stream data for {camera_update_count} match(es)")
+        if point_update_count > 0:
+            update_messages.append(f"Updated camera indices for {point_update_count} point(s)")
+
+    msg = (
+        f"Field updated successfully! {' '.join(update_messages)}."
+        if update_messages
+        else "Field updated successfully!"
+    )
+    db.session.commit()
+    return jsonify({"success": True, "message": msg}), 200
+
+
+@bp.route("/<tournament_url>/delete-field", methods=["POST"])
+@require_tournament_organizer("Only tournament organizers can access this page")
+def delete_field(tournament_url):
+    """Delete field."""
+    field_id = request.form.get("field_id")
+    if not field_id:
+        return jsonify({"success": False, "error": "Field ID is required"}), 400
+
+    field = Field.query.get_or_404(field_id)
+    db.session.delete(field)
+    db.session.commit()
+    return jsonify({"success": True, "message": "Field deleted successfully!"}), 200
+
+
+@bp.route("/<tournament_url>/add-tag", methods=["POST"])
+@require_tournament_organizer("Only tournament organizers can access this page")
+def add_tag(tournament_url):
+    """Add a tag to tournament."""
+    tag = Tag(event=tournament_url, name=request.form["tag_name"])
+
+    db.session.add(tag)
+    db.session.commit()
+
+    return jsonify({"success": True, "message": "Tag added successfully!"}), 200
+
+
+@bp.route("/<tournament_url>/delete-tag", methods=["POST"])
+@require_tournament_organizer("Only tournament organizers can access this page")
+def delete_tag(tournament_url):
+    """Delete tag."""
+    tag_id = request.form.get("tag_id")
+    if not tag_id:
+        return jsonify({"success": False, "error": "Tag ID is required"}), 400
+
+    tag = Tag.query.get_or_404(tag_id)
+    db.session.delete(tag)
+    db.session.commit()
+    return jsonify({"success": True, "message": "Tag deleted successfully!"}), 200
+
+
